chore: Move ShinyModel to stats package
This package should be a bit more fitting place for handling item-related stats.
